# artifacts/openapi_contract_snapshot.json
{
  "components": {
    "schemas": {
...
      "Body_uploadCsv": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_uploadCsv",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobStatus": {
        "enum": [
          "pending",
          "running",
          "succeeded",
          "failed",
          "partial",
          "interrupted"
        ],
        "title": "JobStatus",
        "type": "string"
      },
      "JobType": {
        "description": "What the caller asked for. Determines the stage graph.",
        "enum": [
          "structured_pipeline",
          "document_pipeline",
          "incremental_sync",
          "drift_check",
          "api_spec_preparation",
          "source_native_pipeline",
          "schema_pipeline"
        ],
        "title": "JobType",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "Shared service key for trusted server-to-server integration. Send it as the `X-API-Key` header. Health and documentation routes never require it.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "\nControl plane for the ERP-Aware Data Transformation Pipeline.\n\nThis API **orchestrates** the pipeline's existing phases - schema discovery,\nmapping, transformation, validation, incremental sync, embedding and hybrid\ntiered vector storage. It does not reimplement any of them.\n\n**Boundaries.** Uploaded OpenAPI and Postman documents are parsed as contracts;\nthe endpoints they describe are never called. No LLM is used and no generated\nanswers are produced.\n",
    "title": "ERP-Aware Data Transformation Pipeline API",
    "version": "1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/files/csv": {
      "post": {
        "description": "Upload a CSV and infer its schema through Phase 6.\n\nReturns metadata only. The rows themselves are not echoed: an ingestion\nendpoint that replayed business data would be an accidental data-export\nendpoint.",
        "operationId": "uploadCsv",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadCsv"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsvUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Csv",
        "tags": [
          "files"
        ]
      }
    },
    "/v1/files/documents": {
      "post": {
        "description": "Upload a PDF or image, extract it, and start indexing it.\n\nEvery identity field is OPTIONAL, so a client that posts only a file - which\nis what the existing frontend does - behaves exactly as it did before, plus\nthe indexing that used to require a second call.\n\nIdentity is DECLARED, never inferred. A file named\n``EMP002_birth_certificate.jpg`` uploaded without metadata carries no\nbusiness key: see ``orchestration.document_identity`` for why a guess in\nthat field would be worse than an absence.\n\nReturns metadata only - never the extracted text, which is the document's\nentire content.",
        "operationId": "uploadDocument",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadDocument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Document",
        "tags": [
          "files"
        ]
      }
    },
...
    "/v1/jobs": {
      "get": {
        "operationId": "listJobs",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "job_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job Type"
            }
          },
          {
            "in": "query",
            "name": "source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobResponse"
                  },
                  "title": "Response Listjobs",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "description": "Accept a job and return immediately.\n\n202, not 200: the pipeline runs on the executor afterwards. Blocking here\nuntil embeddings finished would tie a request socket to a model run.",
        "operationId": "createJob",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/v1/representations/{representation_id}": {
      "get": {
        "description": "Resolve a search hit into the AI text it stands for.\n\nThe ``:path`` converter is required, not decorative: representation ids\ncontain colons and dots (``ai:document:erp_legacy_hr_...``), and the\ndefault converter stops at the first slash.\n\nLooks up by ``representation_id`` rather than ``document_id``, and that\nchoice is load-bearing. Two employees issued the same certificate share a\n``document_id``; resolving by it would collapse their association and hand\nback a row that belongs to whichever was written last.",
        "operationId": "getRepresentation",
        "parameters": [
          {
            "in": "path",
            "name": "representation_id",
            "required": true,
            "schema": {
              "title": "Representation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepresentationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Representation",
        "tags": [
          "representations"
        ]
      }
    },
    "/v1/responses/adapt": {
      "post": {
        "description": "Transform an already-executed ERP API response into LLM-ready context.\n\nReturns 200 even when the adaptation only partly succeeded - a refused\nasset URL or a truncating budget is reported in ``warnings`` and\n``partial``, not as an HTTP error, because the fields that DID adapt are\nstill the answer the caller needs. A 422 means the request itself could not\nbe interpreted.",
        "operationId": "adaptResponse",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponseAdaptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseAdaptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Adapt Response",
        "tags": [
          "responses"
        ]
      }
    },
    "/v1/schemas/{schema_id}": {
      "get": {
        "operationId": "getSchema",
        "parameters": [
          {
            "in": "path",
            "name": "schema_id",
            "required": true,
            "schema": {
              "title": "Schema Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Schema",
        "tags": [
          "schemas"
        ]
      }
    },
...
  }
}